#4 chore: reset css 분리
